+++ lib/iex/quote.rb
@@ -15,15 +15,24 @@ module IEX
     property 'latest_source', from: 'latestSource' # the source of latestPrice - IEX real time price, 15 minute delayed price, Close or Previous close
     property 'latest_time', from: 'latestTime' # human readable time of the latestPrice
     property 'latest_update', from: 'latestUpdate' # the update time of latestPrice in milliseconds since midnight Jan 1, 1970
+    property 'latest_update_t', from: 'latestUpdate', with: ->(v) { Time.at(v / 1000) } # the update time of latestPrice
     property 'latest_volume', from: 'latestVolume' # the total market volume of the stock
     property 'iex_realtime_price', from: 'iexRealtimePrice' # last sale price of the stock on IEX
     property 'iex_realtime_size', from: 'iexRealtimeSize' # last sale size of the stock on IEX
     property 'iex_last_updated', from: 'iexLastUpdated' # last update time of the data in milliseconds since midnight Jan 1, 1970 UTC or -1 or 0; if the value is -1 or 0, IEX has not quoted the symbol in the trading day
+    property 'iex_last_updated_t', from: 'iexLastUpdated', with: ->(v) { v > 0 ? Time.at(v / 1000) : nil } # last update time of the data
     property 'delayed_price', from: 'delayedPrice' # 15 minute delayed market price
     property 'delayed_price_time', from: 'delayedPriceTime' # time of the delayed market price
     property 'previous_close', from: 'previousClose' # adjusted close price of the last trading day of the stock
     property 'change' # change in value, calculated using calculation_price from previous_close
     property 'change_percent', from: 'changePercent' # change in percent, calculated using calculation_price from previous_close
+    property 'change_percent_s', from: 'changePercent', with: lambda { |v|
+      [
+        v > 0 ? '+' : '',
+        format('%.2f', v * 100),
+        '%'
+      ].join
+    } # change in percent as a String with a leading + or - sign
     property 'iex_market_percent', from: 'iexMarketPercent' # IEX’s percentage of the market in the stock
     property 'iex_volume', from: 'iexVolume' # shares traded in the stock on IEX
     property 'avg_total_volume', from: 'avgTotalVolume' # 30 day average volume on all markets
@@ -36,14 +45,6 @@ module IEX
     property 'week_52_high', from: 'week52High' # adjusted 52 week high
     property 'week_52_low', from: 'week52Low' # adjusted 52 week low
     property 'ytd_change', from: 'ytdChange' # price change percentage from start of year to previous close
-
-    def change_percent_s
-      [
-        change_percent > 0 ? '+' : '',
-        format('%.2f', change_percent * 100),
-        '%'
-      ].join
-    end
 
     def self.get(symbol)
       new IEX::Api::Quote.get(q: symbol)