+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.2.0 (Next)
 
+* Added `Quote#latest_update_t` and `Quote#iex_last_updated_t` in `Time` format - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2018/03/23)